Camera now zooming out to englobe entire stage
The methos takes the positions of the 4 blocks at the edge of the stage then adds an offSet to that position and check if that 4 positions are inside the viewport.

First the stage is built and then the camera zooms until all satge is visible (inside a contained loop). once that happens, we now know the the ortosize value the camera needs to have to see the entire stage.

After that, in the CameraScript Uptade we keep zooming in and out always trying to match the Stage's Default OrtoSize